+++ packages/block-library/src/navigation/edit/use-convert-classic-menu-to-block-menu.js
@@ -97,7 +97,7 @@ function useConvertClassicToBlockMenu( clientId ) {
 					'wp_navigation',
 					navigationMenu.id,
 					{
-						status: postStatus,
+						status: 'publish',
 					},
 					{ throwOnError: true }
 				);