+++ packages/e2e-test-utils-playwright/src/editor/site-editor.ts
@@ -14,9 +14,10 @@ export async function saveSiteEditorEntities( this: Editor ) {
 	);
 	// Second Save button in the entities panel.
 	await this.page.click(
-		'role=region[name="Save sidebar"i] >> role=button[name="Save"i]'
+		'role=region[name="Save panel"i] >> role=button[name="Save"i]'
 	);
-	await this.page.waitForSelector(
-		'role=region[name="Editor top bar"i] >> role=button[name="Save"i][disabled]'
-	);
+	// A role selector cannot be used here because it needs to check that the `is-busy` class is not present.
+	await this.page.waitForSelector( '[aria-label="Saved"].is-busy', {
+		state: 'hidden',
+	} );
 }