+++ packages/react-native-editor/CHANGELOG.md
@@ -10,6 +10,10 @@ For each user feature we should also add a importance categorization label  to i
 -->
 
 ## Unreleased
+-   [*] Fix crash when trying to convert to regular blocks an undefined/deleted reusable block [#50475]
+
+## 1.94.0
+-   [*] Split pasted content between title and body. [#37169]
 
 ## 1.93.1
 -   [**] Fix regression with the Color hook and ColorPanel. [#49917]